Migrate build to JUnit5 Vintage engine (#12952)
As the prerequisite of dogfooding distributed test.
