+++ buildSrc/subprojects/integration-testing/src/main/kotlin/org/gradle/gradlebuild/test/integrationtests/IntegrationTestsPlugin.kt
@@ -18,7 +18,7 @@ package org.gradle.gradlebuild.test.integrationtests
 import org.gradle.api.Action
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.tasks.testing.junit.JUnitOptions
+import org.gradle.api.tasks.testing.junitplatform.JUnitPlatformOptions
 
 
 class IntegrationTestsPlugin : Plugin<Project> {
@@ -30,7 +30,7 @@ class IntegrationTestsPlugin : Plugin<Project> {
 
         createTestTask("integMultiVersionTest", "forking", sourceSet, TestType.INTEGRATION, Action {
             // This test task runs only multi-version tests and is intended to be used in the late pipeline to sweep up versions not previously tested
-            (options as JUnitOptions).includeCategories("org.gradle.integtests.fixtures.ContextualMultiVersionTest")
+            (options as JUnitPlatformOptions).includeTags("org.gradle.integtests.fixtures.ContextualMultiVersionTest")
         })
     }
 }
