+++ buildSrc/subprojects/plugins/src/main/kotlin/org/gradle/gradlebuild/unittestandcompile/UnitTestAndCompilePlugin.kt
@@ -190,8 +190,10 @@ class UnitTestAndCompilePlugin : Plugin<Project> {
             val implementation = configurations.getByName("implementation")
             val compileOnly = configurations.getByName("compileOnly")
             val testImplementation = configurations.getByName("testImplementation")
+            val testCompileOnly = configurations.getByName("testCompileOnly")
             val testRuntimeOnly = configurations.getByName("testRuntimeOnly")
-            testImplementation(library("junit"))
+            testCompileOnly(library("junit"))
+            testRuntimeOnly(library("junit5_vintage"))
             testImplementation(library("groovy"))
             testImplementation(testLibrary("spock"))
             testRuntimeOnly(testLibrary("bytebuddy"))
@@ -270,6 +272,7 @@ class UnitTestAndCompilePlugin : Plugin<Project> {
         tasks.withType<Test>().configureEach {
             maxParallelForks = project.maxParallelForks
 
+            useJUnitPlatform()
             configureJvmForTest()
             configureGitInfo()
             addOsAsInputs()
