+++ buildSrc/subprojects/plugins/src/main/kotlin/org/gradle/plugins/performance/PerformanceTestPlugin.kt
@@ -11,7 +11,7 @@ import org.gradle.api.tasks.SourceSet
 import org.gradle.api.tasks.TaskCollection
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.api.tasks.bundling.Zip
-import org.gradle.api.tasks.testing.junit.JUnitOptions
+import org.gradle.api.tasks.testing.junitplatform.JUnitPlatformOptions
 import org.gradle.internal.hash.HashUtil
 import org.gradle.kotlin.dsl.*
 import org.gradle.plugins.ide.eclipse.EclipsePlugin
@@ -229,18 +229,18 @@ class PerformanceTestPlugin : Plugin<Project> {
         }
 
         create("performanceTest") {
-            (options as JUnitOptions).apply {
-                includeCategories(performanceRegressionTestCategory)
-                excludeCategories(slowPerformanceRegressionTestCategory)
+            (options as JUnitPlatformOptions).apply {
+                includeTags(performanceRegressionTestCategory)
+                excludeTags(slowPerformanceRegressionTestCategory)
             }
         }
 
         create("slowPerformanceTest") {
-            (options as JUnitOptions).includeCategories(slowPerformanceRegressionTestCategory)
+            (options as JUnitPlatformOptions).includeTags(slowPerformanceRegressionTestCategory)
         }
 
         create("performanceExperiment") {
-            (options as JUnitOptions).includeCategories(performanceExperimentCategory)
+            (options as JUnitPlatformOptions).includeTags(performanceExperimentCategory)
         }
 
         create("fullPerformanceTest")
@@ -263,31 +263,31 @@ class PerformanceTestPlugin : Plugin<Project> {
         }
 
         create("distributedPerformanceTest", DistributedPerformanceTest::class) {
-            (options as JUnitOptions).apply {
-                includeCategories(performanceRegressionTestCategory)
-                excludeCategories(slowPerformanceRegressionTestCategory)
+            (options as JUnitPlatformOptions).apply {
+                includeTags(performanceRegressionTestCategory)
+                excludeTags(slowPerformanceRegressionTestCategory)
             }
             channel = "commits"
             retryFailedScenarios()
         }
         create("distributedSlowPerformanceTest", DistributedPerformanceTest::class) {
-            (options as JUnitOptions).includeCategories(slowPerformanceRegressionTestCategory)
+            (options as JUnitPlatformOptions).includeTags(slowPerformanceRegressionTestCategory)
             channel = "commits"
             retryFailedScenarios()
         }
         create("distributedPerformanceExperiment", DistributedPerformanceTest::class) {
-            (options as JUnitOptions).includeCategories(performanceExperimentCategory)
+            (options as JUnitPlatformOptions).includeTags(performanceExperimentCategory)
             channel = "experiments"
             retryFailedScenarios()
         }
         create("distributedHistoricalPerformanceTest", DistributedPerformanceTest::class) {
-            (options as JUnitOptions).excludeCategories(performanceExperimentCategory)
+            (options as JUnitPlatformOptions).excludeTags(performanceExperimentCategory)
             configuredBaselines.set(Config.baseLineList)
             checks = "none"
             channel = "historical"
         }
         create("distributedFlakinessDetection", DistributedPerformanceTest::class) {
-            (options as JUnitOptions).includeCategories(performanceRegressionTestCategory)
+            (options as JUnitPlatformOptions).includeTags(performanceRegressionTestCategory)
             distributedPerformanceReporter.reportGeneratorClass = "org.gradle.performance.results.report.FlakinessReportGenerator"
             repeatScenarios(3)
             checks = "none"
@@ -453,6 +453,8 @@ class PerformanceTestPlugin : Plugin<Project> {
             binaryDistributions.binZipRequired = true
             libsRepository.required = true
             maxParallelForks = 1
+
+            useJUnitPlatform()
 
             project.findProperty(PropertyNames.baselines)?.let { baselines ->
                 task.configuredBaselines.set(baselines as String)
