+++ gradle/dependencies.gradle
@@ -71,8 +71,8 @@ libraries.joda =                [coordinates: 'joda-time:joda-time', version: '2
 libraries.jsch =                [coordinates: 'com.jcraft:jsch', version: '0.1.55', license: 'BSD-style']
 libraries.jsr305 =              [coordinates: 'com.google.code.findbugs:jsr305', version: '3.0.2', license: "3-Clause BSD"]
 libraries.junit =               [coordinates: 'junit:junit', version: '4.13', license: "Eclipse Public License 1.0"]
-libraries.junit_platform =      [coordinates: 'org.junit.platform:junit-platform-launcher', version: '1.6.0', license: "Eclipse Public License 2.0"]
-libraries.junit_api =           [coordinates: 'org.junit.jupiter:junit-jupiter-api', version: '5.6.0', license: "Eclipse Public License 2.0"]
+libraries.junit_platform =      [coordinates: 'org.junit.platform:junit-platform-launcher', version: '1.6.2', license: "Eclipse Public License 2.0"]
+libraries.junit5_vintage =     [coordinates: 'org.junit.vintage:junit-vintage-engine', version: '5.6.2', license: "Eclipse Public License 2.0"]
 libraries.kryo =                [coordinates: 'com.esotericsoftware.kryo:kryo', version: '2.24.0', license: "3-Clause BSD"]
 libraries.maven3 =              [coordinates: 'org.apache.maven:maven-core', version: '3.0.5', license: "Apache 2.0"]
 libraries.maven3_wagon_file =   [coordinates: 'org.apache.maven.wagon:wagon-file', version: '3.0.0', because: '3.1.0 of wagon-http seems to break Digest authentication', license: "Apache 2.0"]
