+++ subprojects/performance/src/performanceTest/groovy/org/gradle/performance/regression/android/RealLifeAndroidBuildPerformanceTest.groovy
@@ -16,33 +16,14 @@
 
 package org.gradle.performance.regression.android
 
-import org.gradle.integtests.fixtures.versions.AndroidGradlePluginVersions
-import org.gradle.internal.scan.config.fixtures.GradleEnterprisePluginSettingsFixture
-import org.gradle.performance.AbstractCrossVersionGradleProfilerPerformanceTest
-import org.gradle.performance.categories.SlowPerformanceRegressionTest
-import org.gradle.profiler.BuildMutator
-import org.gradle.profiler.ScenarioContext
-import org.gradle.profiler.mutations.AbstractCleanupMutator
-import org.gradle.profiler.mutations.ClearArtifactTransformCacheMutator
-import org.junit.experimental.categories.Category
+
 import spock.lang.Unroll
 
 import static org.gradle.performance.regression.android.AndroidTestProject.K9_ANDROID
 import static org.gradle.performance.regression.android.AndroidTestProject.LARGE_ANDROID_BUILD
 import static org.gradle.performance.regression.android.IncrementalAndroidTestProject.SANTA_TRACKER_KOTLIN
 
-class RealLifeAndroidBuildPerformanceTest extends AbstractCrossVersionGradleProfilerPerformanceTest {
-
-    private static final String SANTA_AGP_TARGET_VERSION = "3.6"
-
-    def setup() {
-        runner.args = [AndroidGradlePluginVersions.OVERRIDE_VERSION_CHECK]
-        runner.targetVersions = ["6.5-20200512220058+0000"]
-        // AGP 3.6 requires 5.6.1+
-        // The enterprise plugin requires Gradle 6.0
-        runner.minimumBaseVersion = "6.0"
-    }
-
+class RealLifeAndroidBuildPerformanceTest extends AbstractRealLifeAndroidBuildPerformanceTest {
     @Unroll
     def "#tasks on #testProject"() {
         given:
@@ -75,39 +56,6 @@ class RealLifeAndroidBuildPerformanceTest extends AbstractCrossVersionGradleProf
         LARGE_ANDROID_BUILD  | true     | null       | null | 'assembleDebug'
         LARGE_ANDROID_BUILD  | true     | 2          | 8    | 'clean phthalic:assembleDebug'
         SANTA_TRACKER_KOTLIN | true     | null       | null | 'assembleDebug'
-    }
-
-    @Category(SlowPerformanceRegressionTest)
-    @Unroll
-    def "clean #tasks on #testProject with clean transforms cache"() {
-        given:
-        testProject.configure(runner)
-        runner.tasksToRun = tasks.split(' ')
-        runner.args.add('-Dorg.gradle.parallel=true')
-        runner.warmUpRuns = warmUpRuns
-        runner.cleanTasks = ["clean"]
-        runner.runs = runs
-        runner.addBuildMutator { invocationSettings ->
-            new ClearArtifactTransformCacheMutator(invocationSettings.getGradleUserHome(), AbstractCleanupMutator.CleanupSchedule.BUILD)
-        }
-        applyEnterprisePlugin()
-
-        and:
-        if (testProject == SANTA_TRACKER_KOTLIN) {
-            (testProject as IncrementalAndroidTestProject).configureForLatestAgpVersionOfMinor(runner, SANTA_AGP_TARGET_VERSION)
-        }
-
-        when:
-        def result = runner.run()
-
-        then:
-        result.assertCurrentVersionHasNotRegressed()
-
-        where:
-        testProject          | warmUpRuns | runs | tasks
-        LARGE_ANDROID_BUILD  | 2          | 8    | 'phthalic:assembleDebug'
-        LARGE_ANDROID_BUILD  | 2          | 8    | 'assembleDebug'
-        SANTA_TRACKER_KOTLIN | null       | null | 'assembleDebug'
     }
 
     @Unroll
@@ -144,16 +92,5 @@ class RealLifeAndroidBuildPerformanceTest extends AbstractCrossVersionGradleProf
 
         where:
         testProject << [SANTA_TRACKER_KOTLIN]
-    }
-
-    void applyEnterprisePlugin() {
-        runner.addBuildMutator { invocationSettings ->
-            new BuildMutator() {
-                @Override
-                void beforeScenario(ScenarioContext context) {
-                    GradleEnterprisePluginSettingsFixture.applyEnterprisePlugin(new File(invocationSettings.projectDir, "settings.gradle"))
-                }
-            }
-        }
     }
 }
