+++ subprojects/soak/soak.gradle.kts
@@ -38,8 +38,8 @@ dependencies {
 
 tasks.integTest {
     options {
-        require(this is JUnitOptions)
-        excludeCategories("org.gradle.soak.categories.SoakTest")
+        require(this is JUnitPlatformOptions)
+        excludeTags("org.gradle.soak.categories.SoakTest")
     }
 }
 
@@ -49,8 +49,8 @@ tasks.register("soakIntegTest", org.gradle.gradlebuild.test.integrationtests.Soa
     classpath = integTestSourceSet.runtimeClasspath
     systemProperty("org.gradle.soaktest", "true")
     options {
-        require(this is JUnitOptions)
-        includeCategories("org.gradle.soak.categories.SoakTest")
+        require(this is JUnitPlatformOptions)
+        includeTags("org.gradle.soak.categories.SoakTest")
     }
 }
