+++ subprojects/testing-jvm/testing-jvm.gradle.kts
@@ -79,6 +79,8 @@ classycle {
 tasks.named<Test>("test").configure {
     exclude("org/gradle/api/internal/tasks/testing/junit/ATestClass*.*")
     exclude("org/gradle/api/internal/tasks/testing/junit/ABroken*TestClass*.*")
+    exclude("org/gradle/api/internal/tasks/testing/junit/ATestSetUpWithBrokenSetUp*.*")
+    exclude("org/gradle/api/internal/tasks/testing/testng/ATestNGFactoryClass*.*")
 }
 
 integrationTestUsesSampleDir("subprojects/testing-jvm/src/main")
